fix server memory leak (#1524)
After v24 release, the server had has a bad memory leak, causing worker
memory to rise to 1gb after ~12 hours.

I believed this is caused by not removing listeners on websocket when
kicking a client.

- [x] I have added screenshots for all UI updates
- [x] I process any text displayed to the user through translateText()
and I've added it to the en.json file
- [x] I have added relevant tests to the test directory
- [x] I confirm I have thoroughly tested these changes and take full
responsibility for any bugs introduced
- [x] I have read and accepted the CLA aggreement (only required once).
regression is found:

evan